added an accessor to clear the body on a geometry.  Added a method to get the inertia on a generic body, and added methods to enable and disable gravity.
Implemented to IsDynamic, IsKinematic, and IsStatic methods on the generic body so they don't have to be implemented in every engine.

Added the pal Character interface and implemented it for bullet.

Implemented the generic body for ODE, and it mostly works.  It uses the built in ode inertia calls, but it seems like it may be better to use the pal built-ins for better consistency.

Fixed a memory leak and crash on exit for bullet, also made it use the pal code to set the inertia because it works better than the bullet compound body code, which just uses the AABB.