Added a new sub project, the "PAL Benchmark " to SVN control.
The PAL Benchmark runs a GUI using SDL and Irrlicht to render benchmark tests of most of the physics engines that PAL supports.

CMake is not required to build the project, however at the moment only Visual Studio on Windows 7 (and probably other versions) is running. I am working on a version that runs on Mac OS X Snow Leopard using Xcode 3.2.